Move proof logic away from viper
