+++ client/commands/query/get.go
@@ -10,12 +10,10 @@ import (
 
 	wire "github.com/tendermint/go-wire"
 	"github.com/tendermint/go-wire/data"
-	lc "github.com/tendermint/light-client"
-	"github.com/tendermint/light-client/certifiers"
 	"github.com/tendermint/light-client/proofs"
 	"github.com/tendermint/merkleeyes/iavl"
-	"github.com/tendermint/tendermint/rpc/client"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/commands"
 )
 
@@ -67,98 +65,7 @@ func GetWithProof(key []byte) (data.Bytes, uint64,
 	if err != nil {
 		return nil, 0, nil, nil, err
 	}
-	return getWithProof(key, node, cert)
-}
-
-func getWithProof(key []byte, node client.Client, cert certifiers.Certifier) (
-	val data.Bytes, height uint64, eproof *iavl.KeyExistsProof, neproof *iavl.KeyNotExistsProof, err error) {
-
-	resp, err := node.ABCIQuery("/key", key, true)
-	if err != nil {
-		return
-	}
-
-	// make sure the proof is the proper height
-	if !resp.Code.IsOK() {
-		err = errors.Errorf("Query error %d: %s", resp.Code, resp.Code.String())
-		return
-	}
-	if len(resp.Key) == 0 || len(resp.Proof) == 0 {
-		err = lc.ErrNoData()
-		return
-	}
-	if resp.Height == 0 {
-		err = errors.New("Height returned is zero")
-		return
-	}
-
-	check, err := getCertifiedCheckpoint(int(resp.Height), node, cert)
-	if err != nil {
-		return
-	}
-
-	if len(resp.Value) > 0 {
-		// The key was found, construct a proof of existence.
-		eproof = new(iavl.KeyExistsProof)
-		err = wire.ReadBinaryBytes(resp.Proof, &eproof)
-		if err != nil {
-			err = errors.Wrap(err, "Error reading proof")
-			return
-		}
-
-		// Validate the proof against the certified header to ensure data integrity.
-		err = eproof.Verify(resp.Key, resp.Value, check.Header.AppHash)
-		if err != nil {
-			err = errors.Wrap(err, "Couldn't verify proof")
-			return
-		}
-		val = data.Bytes(resp.Value)
-	} else {
-		// The key wasn't found, construct a proof of non-existence.
-		neproof = new(iavl.KeyNotExistsProof)
-		err = wire.ReadBinaryBytes(resp.Proof, &neproof)
-		if err != nil {
-			err = errors.Wrap(err, "Error reading proof")
-			return
-		}
-
-		// Validate the proof against the certified header to ensure data integrity.
-		err = neproof.Verify(resp.Key, check.Header.AppHash)
-		if err != nil {
-			err = errors.Wrap(err, "Couldn't verify proof")
-			return
-		}
-		err = lc.ErrNoData()
-	}
-
-	height = resp.Height
-	return
-}
-
-// getCertifiedCheckpoint gets the signed header for a given height
-// and certifies it.  Returns error if unable to get a proven header.
-func getCertifiedCheckpoint(h int, node client.Client,
-	cert certifiers.Certifier) (empty lc.Checkpoint, err error) {
-
-	// FIXME: cannot use cert.GetByHeight for now, as it also requires
-	// Validators and will fail on querying tendermint for non-current height.
-	// When this is supported, we should use it instead...
-	client.WaitForHeight(node, h, nil)
-	commit, err := node.Commit(h)
-	if err != nil {
-		return
-	}
-	check := lc.Checkpoint{
-		Header: commit.Header,
-		Commit: commit.Commit,
-	}
-
-	// validate downloaded checkpoint with our request and trust store.
-	if check.Height() != h {
-		return empty, lc.ErrHeightMismatch(h, check.Height())
-	}
-	err = cert.Certify(check)
-	return check, nil
+	return client.GetWithProof(key, node, cert)
 }
 
 // ParseHexKey parses the key flag as hex and converts to bytes or returns error
