+++ client/common.go
@@ -46,3 +46,9 @@ func GetCertifier(chainID string, trust certifiers.Provider,
 	cert := certifiers.NewInquiring(chainID, seed, trust, source)
 	return cert, nil
 }
+
+// GetSecureNode uses a given certifier to wrap an connection to an untrusted
+// host and return a cryptographically secure rpc client.
+func GetSecureNode(c rpcclient.Client, cert *certifiers.InquiringCertifier) rpcclient.Client {
+	return certclient.Wrap(c, cert)
+}
